fixed:调整v2.5.0测试版bug (#931)
* update readme badge for en

* server/middleware/operation.go : fix 当记录日志当请求为GET时，不对query参数进行记录

* update element badge to element--plus-2.0.1-green

* fix unmatched type of id field

* update badge for vue gorm in readme

* fixed:#917

+++ server/service/system/sys_api.go
@@ -115,7 +115,7 @@ func (apiService *ApiService) GetAllApis() (err error, apis []system.SysApi) {
 //@param: id float64
 //@return: err error, api model.SysApi
 
-func (apiService *ApiService) GetApiById(id float64) (err error, api system.SysApi) {
+func (apiService *ApiService) GetApiById(id int) (err error, api system.SysApi) {
 	err = global.GVA_DB.Where("id = ?", id).First(&api).Error
 	return
 }
